requirements.txt
crewai[tools]>=0.152.0
fastapi>=0.116.2
uvicorn>=0.32.0
playwright>=1.55.0
python-dotenv>=1.1.1
pydantic>=2.0.0